requirements.txt
python-dotenv
py-cord
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
lxml
bs4
email_reply_parser